refactor: change name folder as dashboardAdmin
