+++ pages/admin/index.jsx
@@ -1,6 +1,6 @@
 import { Box } from "@chakra-ui/react";
 import { NavBar } from "../../components/admin/navbar/NavBar";
-import { DashboardAdmin } from "../../components/admin/listEvents/DashboardAdmin";
+import { DashboardAdmin } from "../../components/admin/dashboardAdmin/DashboardAdmin";
 
 const index = () => {
   return (
